Crate llhd

The Low Level Hardware Description language. This library provides tools to create, modify, store, and load LLHD graphs.

Re-exports

pub use self::TypeKind::*;

Modules

assembly
Facilities to emit a module as human-readable assembly, or to parse such assembly back into a module.

ir
Representation of LLHD functions, processes, and entitites.

pass
Optimization and analysis passes on LLHD IR.

table
Primary and secondary tables.

util
Various utility functions that fit nowhere else.

verifier
Verification of IR integrity.

Macros

impl_table_indexing
Generate the Index and IndexMut operations for a contained table.

impl_table_key
Generate a new opaque table key struct.

Structs

ConstTime
A constant time value.

Enums

TypeKind
The different kinds of types.

Functions

array_ty
Create a array type. size is the number of elements in the array, and ty the type of each individual element.

entity_ty
Create an entity type with the given input and output arguments.

enum_ty
Create an enum type of the requested size.

func_ty
Create a function type with the given arguments and return type.

int_ty
Create an integer type of the requested size.

pointer_ty
Create a pointer type with the requested data type.

signal_ty
Create a signal type with the requested data type.

struct_ty
Create a struct type. fields is an list of types, one for each field.

time_ty
Create a time type.

void_ty
Create a void type.

Type Definitions

Type