Enum llhd::ValueRef

pub enum ValueRef {
    Inst(InstRef),
    Block(BlockRef),
    Argument(ArgumentRef),
    Function(FunctionRef),
    Process(ProcessRef),
    Entity(EntityRef),
    Global,
    Const(Const),
    Aggregate(Aggregate),
}

A reference to a value in a module.

Variants
Inst(InstRef)Block(BlockRef)Argument(ArgumentRef)Function(FunctionRef)Process(ProcessRef)Entity(EntityRef)GlobalConst(Const)Aggregate(Aggregate)
Methods

impl ValueRef

pub fn into_const(self) -> Const

Convert this value reference into the constant it contains. Panics if the value reference does not contain a constant.

pub fn as_const(&self) -> &Const

Unwrap and return a reference to the constant represented by this value reference. Panics if the value reference does not contain a constant.

pub fn maybe_const(&self) -> Option<&Const>

Try to unwrap this value reference as a constant.

pub fn id(&self) -> Option<ValueId>

Obtain the ID of the value this reference points to, or None if the value has no ID (e.g. if it is a constant).

pub fn unwrap_inst(&self) -> InstRef

Unwrap this reference as an instruction.

Panics if this is not an instruction.

pub fn unwrap_block(&self) -> BlockRef

Unwrap this reference as a block.

Panics if this is not a block.

pub fn unwrap_argument(&self) -> ArgumentRef

Unwrap this reference as an argument.

Panics if this is not an argument.

pub fn unwrap_function(&self) -> FunctionRef

Unwrap this reference as a function.

Panics if this is not a function.

pub fn unwrap_process(&self) -> ProcessRef

Unwrap this reference as a process.

Panics if this is not a process.

pub fn unwrap_entity(&self) -> EntityRef

Unwrap this reference as an entity.

Panics if this is not an entity.

pub fn unwrap_const(&self) -> &Const

Unwrap this reference as a constant.

Panics if this is not a constant.

pub fn unwrap_aggregate(&self) -> &Aggregate

Unwrap this reference as an aggregate.

Panics if this is not an aggregate.