Module llhd::ir

Representation of LLHD functions, processes, and entitites.

This module implements the intermediate representation around which the rest of the framework is built.

Modules

prelude
Re-exports of commonly used IR items.

Structs

Arg
An argument of a Function, Process, or Entity.

Block
A basic block.

BlockData
Internal table storage for blocks.

ControlFlowGraph
A control flow graph.

DataFlowGraph
A data flow graph.

Entity
An entity.

EntityBuilder
Temporary object used to build a single Entity.

ExtUnit
An external Function, Process or Entity.

ExtUnitData
Another unit referenced within a Function, Process, or Entity.

Function
A function.

FunctionBuilder
Temporary object used to build a single Function.

FunctionLayout
Determines the order of instructions and BBs in a Function or Process.

Inst
An instruction.

InstBuilder
A temporary object used to construct a single instruction.

InstDumper
Temporary object to dump an Inst in human-readable form for debugging.

InstLayout
Determines the order of instructions.

ModUnit
An unit definition or declaration in a module.

Module
A module.

ModuleDumper
Temporary object to dump a Module in human-readable form for debugging.

Process
A process.

ProcessBuilder
Temporary object used to build a single Process.

Signature
A description of the input and output arguments of a unit.

SignatureDumper
Temporary object to dump a Signature in human-readable form for debugging.

UnitDumper
Temporary object to dump an Entity in human-readable form for debugging.

UnitFlags
A set of flags identifying a unit.

Value
A value.

Enums

AnyObject
Any one of the table keys in this module.

InstData
An instruction format.

ModUnitData
Internal table storage for units in a module.

Opcode
An instruction opcode.

RegMode
The trigger modes for register data acquisition.

UnitKind
The three different units that may appear in LLHD IR.

UnitName
A name of a function, process, or entity.

ValueData
Internal table storage for values.

Traits

Unit
A Function, Process, or Entity.

UnitBuilder
A temporary object used to populate a Function, Process or Entity.