Enum llhd::ir::ModUnitData

pub enum ModUnitData {
    Function(Function),
    Process(Process),
    Entity(Entity),
    Declare {
        sig: Signature,
        name: UnitName,
    },
}

Internal table storage for units in a module.

Variants

Function(Function)

The unit is a function.

Process(Process)

The unit is a process.

Entity(Entity)

The unit is an entity.

Declare

The unit is a declaration of an external unit.

Fields of Declare

sig: Signaturename: UnitName

Methods

impl ModUnitData

pub fn get_function(&self) -> Option<&Function>

If this unit is a function, return it. Otherwise return None.

pub fn get_function_mut(&mut self) -> Option<&mut Function>

If this unit is a function, return it. Otherwise return None.

pub fn get_process(&self) -> Option<&Process>

If this unit is a process, return it. Otherwise return None.

pub fn get_process_mut(&mut self) -> Option<&mut Process>

If this unit is a process, return it. Otherwise return None.

pub fn get_entity(&self) -> Option<&Entity>

If this unit is an entity, return it. Otherwise return None.

pub fn get_entity_mut(&mut self) -> Option<&mut Entity>

If this unit is an entity, return it. Otherwise return None.

pub fn get_declaration(&self) -> Option<(&UnitName, &Signature)>

If this unit is an external declaration, return it. Otherwise return None.

pub fn get_declaration_mut(&mut self) -> Option<(&mut UnitName, &mut Signature)>

If this unit is an external declaration, return it. Otherwise return None.

pub fn is_function(&self) -> bool

Check whether this is a function.

pub fn is_process(&self) -> bool

Check whether this is a process.

pub fn is_entity(&self) -> bool

Check whether this is an entity.

pub fn is_declaration(&self) -> bool

Check whether this is a declaration of an external unit.

pub fn sig(&self) -> &Signature

Return the signature of the unit.

pub fn name(&self) -> &UnitName

Return the name of the unit.

pub fn get_dfg(&self) -> Option<&DataFlowGraph>

Return the data flow graph of the unit, if there is one.

pub fn get_dfg_mut(&mut self) -> Option<&mut DataFlowGraph>

Return the mutable data flow graph of the unit, if there is one.