Enum llhd::ir::Opcode

pub enum Opcode {
    ConstInt,
    ConstTime,
    Alias,
    ArrayUniform,
    Array,
    Struct,
    Not,
    Neg,
    Add,
    Sub,
    And,
    Or,
    Xor,
    Smul,
    Sdiv,
    Smod,
    Srem,
    Umul,
    Udiv,
    Umod,
    Urem,
    Eq,
    Neq,
    Slt,
    Sgt,
    Sle,
    Sge,
    Ult,
    Ugt,
    Ule,
    Uge,
    Shl,
    Shr,
    Mux,
    Reg,
    InsField,
    InsSlice,
    ExtField,
    ExtSlice,
    Con,
    Del,
    Call,
    Inst,
    Sig,
    Prb,
    Drv,
    Var,
    Ld,
    St,
    Halt,
    Ret,
    RetValue,
    Br,
    BrCond,
    Wait,
    WaitTime,
}

An instruction opcode.

This enum represents the actual instruction, whereas InstData covers the format and arguments of the instruction.

Variants

ConstInt

ConstTime

Alias

ArrayUniform

Array

Struct

Not

Neg

Add

Sub

And

Or

Xor

Smul

Sdiv

Smod

Srem

Umul

Udiv

Umod

Urem

Eq

Neq

Slt

Sgt

Sle

Sge

Ult

Ugt

Ule

Uge

Shl

Shr

Mux

Reg

InsField

InsSlice

ExtField

ExtSlice

Con

Del

Call

Inst

Sig

Prb

Drv

Var

Ld

St

Halt

Ret

RetValue

Br

BrCond

Wait

WaitTime

Methods

impl Opcode

pub fn valid_in(self) -> UnitFlags

Return a set of flags where this instruction is valid.

pub fn valid_in_function(self) -> bool

Check if this instruction can appear in a Function.

pub fn valid_in_process(self) -> bool

Check if this instruction can appear in a Process.

pub fn valid_in_entity(self) -> bool

Check if this instruction can appear in a Entity.

pub fn is_const(self) -> bool

Check if this instruction is a constant.

pub fn is_terminator(self) -> bool

Check if this instruction is a terminator.

pub fn is_return(self) -> bool

Check if this is a return instruction.