Module llhd::ir::prelude

Re-exports of commonly used IR items.

Re-exports

pub use crate::ir::Arg;

pub use crate::ir::Block;

pub use crate::ir::Inst;

pub use crate::ir::Value;

Structs

Entity
An entity.

EntityBuilder
Temporary object used to build a single Entity.

Function
A function.

FunctionBuilder
Temporary object used to build a single Function.

ModUnit
An unit definition or declaration in a module.

Module
A module.

Process
A process.

ProcessBuilder
Temporary object used to build a single Process.

Signature
A description of the input and output arguments of a unit.

Enums

Opcode
An instruction opcode.

RegMode
The trigger modes for register data acquisition.

UnitName
A name of a function, process, or entity.

Traits

Unit
A Function, Process, or Entity.

UnitBuilder
A temporary object used to populate a Function, Process or Entity.