Struct llhd::ir::DataFlowGraph

pub struct DataFlowGraph { /* fields omitted */ }

A data flow graph.

This is the main container for instructions, values, and the relationship between them. Every Function, Process, and Entity has an associated data flow graph.

Methods

impl DataFlowGraph

pub fn new() -> Self

Create a new data flow graph.

pub fn add_placeholder(&mut self, ty: Type) -> Value

Add a placeholder value.

This function is intended to be used when constructing PHI nodes.

pub fn remove_placeholder(&mut self, value: Value)

Remove a placeholder value.

pub fn is_placeholder(&self, value: Value) -> bool

Check if a value is a placeholder.

pub fn add_inst(&mut self, data: InstData, ty: Type) -> Inst

Add an instruction.

pub fn remove_inst(&mut self, inst: Inst)

Remove an instruction.

pub fn has_result(&self, inst: Inst) -> bool

Returns whether an instruction produces a result.

pub fn inst_result(&self, inst: Inst) -> Value

Returns the result of an instruction.

pub fn arg_value(&self, arg: Arg) -> Value

Returns the value of an argument.

pub fn value_type(&self, value: Value) -> Type

Returns the type of a value.

pub fn inst_type(&self, inst: Inst) -> Type

Returns the type of an instruction.

pub fn get_value_inst(&self, value: Value) -> Option<Inst>

Return the instruction that produces value.

pub fn value_inst(&self, value: Value) -> Inst

Return the instruction that produces value, or panic.

pub fn get_name(&self, value: Value) -> Option<&str>

Return the name of a value.

pub fn set_name(&mut self, value: Value, name: String)

Set the name of a value.

pub fn clear_name(&mut self, value: Value) -> Option<String>

Clear the name of a value.

pub fn replace_use(&mut self, from: Value, to: Value) -> usize

Replace all uses of a value with another.

Returns how many uses were replaced.

pub fn uses(&self, value: Value) -> impl Iterator<Item = (Inst, usize)>

Iterate over all uses of a value.

pub fn has_uses(&self, value: Value) -> bool

Check if a value is used.

pub fn has_one_use(&self, value: Value) -> bool

Check if a value has exactly one use.

pub fn replace_block_use(&mut self, from: Block, to: Block) -> usize

Replace all uses of a block with another.

Returns how many blocks were replaced.