Trait llhd::Value

pub trait Value {
    fn id(&self) -> ValueId;

    fn ty(&self) -> Type;

    fn name(&self) -> Option<&str> { ... }

    fn is_global(&self) -> bool { ... }
}

Required methods

fn id(&self) -> ValueId

Get the unique ID of the value.

fn ty(&self) -> Type

Get the type of the value.

Loading content...
Provided methods

fn name(&self) -> Option<&str>

Get the optional name of the value.

fn is_global(&self) -> bool

Whether this value is global or not. Global values are considered during linking, and are visible in a module's symbol table. Local values are not, and are only visible within the surrounding context (module or unit).

Loading content...